pyproject.toml
[tool.poetry]
name = "fastapicodes"
version = "0.1.0"
description = ""
authors = ["Link13-tech <151656048+Link13-tech@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.115.0"}
uvicorn = "^0.30.6"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
httpx = "^0.27.2"
python-jose = {extras = ["cryptogrphy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.9.2"}
python-multipart = "^0.0.10"
bcrypt = "<4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
